requirements.txt
altair==5.5.0
attrs==25.3.0
beautifulsoup4==4.13.5
blinker==1.9.0
cachetools==6.2.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
et_xmlfile==2.0.0
gitdb==4.0.12
GitPython==3.1.45
h11==0.16.0
idna==3.10
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
MarkupSafe==3.0.2
narwhals==2.3.0
numpy==2.3.2
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==25.0
pandas==2.3.2
pillow==11.3.0
protobuf==6.32.0
pyarrow==21.0.0
pycparser==2.22
pydeck==0.9.1
PySocks==1.7.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
referencing==0.36.2
requests==2.32.5
rpds-py==0.27.1
selenium==4.35.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8
streamlit==1.49.1
tenacity==9.1.2
toml==0.10.2
tornado==6.5.2
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
watchdog==6.0.0
webdriver-manager==4.0.2
websocket-client==1.8.0
wsproto==1.2.0
